# src/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Это API для управления пользователями, товарами и поставщиками.",
        "title": "API документация",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/supplier/create": {
            "post": {
                "description": "Создает нового поставщика с указанным адресом",
                "tags": [
                    "suppliers"
                ],
                "summary": "Создать поставщика",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/response.CreateSupplier"
                            }
                        }
                    },
                    "description": "Данные нового поставщика",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Поставщик успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Не удалось создать поставщика",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/supplier/delete/{id}": {
            "delete": {
                "description": "Удаляет поставщика по его ID",
                "tags": [
                    "suppliers"
                ],
                "summary": "Удалить поставщика",
                "parameters": [
                    {
                        "description": "UUID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Поставщик успешно удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID поставщика",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при удалении поставщика",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/supplier/supplierList": {
            "get": {
                "description": "Возвращает список всех поставщиков",
                "tags": [
                    "suppliers"
                ],
                "summary": "Получить список поставщиков",
                "responses": {
                    "200": {
                        "description": "Список поставщиков",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/response.SupplierResponse"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при получении списка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/supplier/updateAddress/{id}": {
            "put": {
                "description": "Обновляет адрес поставщика по его ID",
                "tags": [
                    "suppliers"
                ],
                "summary": "Обновить адрес поставщика",
                "parameters": [
                    {
                        "description": "UUID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/response.CreateUpdateAddress"
                            }
                        }
                    },
                    "description": "Новый адрес поставщика",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Адрес успешно изменен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в данных или некорректный UUID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при обновлении адреса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/supplier/{id}": {
            "get": {
                "description": "Возвращает данные поставщика по его ID",
                "tags": [
                    "suppliers"
                ],
                "summary": "Получить поставщика",
                "parameters": [
                    {
                        "description": "UUID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные поставщика",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.SupplierResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID или отсутствует ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка при получении поставщика",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:5000/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "response.CreateSupplier": {
                "type": "object",
                "properties": {
                    "address": {
                        "$ref": "#/components/schemas/response.CreateUpdateAddress"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    }
                }
            },
            "response.CreateUpdateAddress": {
                "type": "object",
                "properties": {
                    "city": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "street": {
                        "type": "string"
                    }
                }
            },
...
            "response.SupplierResponse": {
                "type": "object",
                "properties": {
                    "address": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}